# [NVIDIA] Support SiluMul + NVFP4 quant fusion