--- a/vllm/model_executor/layers/quantization/modelopt.py
+++ b/vllm/model_executor/layers/quantization/modelopt.py
@@ -885,7 +885,7 @@ class ModelOptNvFp4LinearMethod(LinearMethodBase):
         layer.alpha = Parameter(layer.input_scale * layer.weight_scale_2,
                                 requires_grad=False)
 
-        # Calcuate `1 / input_scale` so that we do not need to do so at runtime.
+        # Calculate `1 / input_scale` so that we don't need to do so at runtime
         layer.input_scale_inv = Parameter(
             (1 / layer.input_scale).to(torch.float32), requires_grad=False)