… .buildkite/test-pipeline.yaml
… CMakeLists.txt
… tests/compile/test_silu_mul_quant_fusion.py
… tests/kernels/quantization/test_silu_nvfp4_quant_fusion.py
… vllm/compilation/fusion.py