<!-- markdownlint-disable -->

## Purpose
Support Silu_Mul + NVFP4 quant fusion(following up #22448).

## Test Plan && Test Result
**Kernel functional:**
`tests/kernels/quantization/test_silu_nvfp4_quant_fusion.py`
```
====== 8 passed in 2.76s =====
```

**Fusion unit test:**
`tests/compile/test_silu_mul_quant_fusion.py`
```
====== 3 passed, 1 skipped, 5 warnings in 4.38s =====
```

**lm_eval && benchmarking:**
main:
```
============ Serving Benchmark Result ============
Successful requests:                     640
Maximum request concurrency:             128
Benchmark duration (s):                  174.31
Total input tokens:                      654052
Total generated tokens:                  655360
Request throughput (req/s):              3.67
Output token throughput (tok/s):         3759.83
Total Token throughput (tok/s):          7512.16
---------------Time to First Token----------------
Mean TTFT (ms):                          1458.45
Median TTFT (ms):                        1067.94
P99 TTFT (ms):                           5477.33
-----Time per Output Token (excl. 1st token)------
Mean TPOT (ms):                          32.61
Median TPOT (ms):                        32.82
P99 TPOT (ms):                           34.01
---------------Inter-token Latency----------------
Mean ITL (ms):                           32.61
Median ITL (ms):                         29.03
P99 ITL (ms):                            331.72
==================================================
vllm ({'pretrained': 'nvidia/Llama-3.3-70B-Instruct-FP4', 'kv_cache_dtype': 'fp8', 'tensor_parallel_size': 1, 'max_model_len': 2048, 'trust_remote_code': True}), gen_kwargs: (temperature=0.0), limit: 500.0, num_fewshot: 5, batch_size: 200
|Tasks|Version|     Filter     |n-shot|  Metric   |   |Value|   |Stderr|
|-----|------:|----------------|-----:|-----------|---|----:|---|-----:|
|gsm8k|      3|flexible-extract|     5|exact_match|↑  |0.934|±  |0.0111|
|     |       |strict-match    |     5|exact_match|↑  |0.858|±  |0.0156|
```
PR:
```
disable fusion:
============ Serving Benchmark Result ============
Successful requests:                     640
Maximum request concurrency:             128
Benchmark duration (s):                  172.20
Total input tokens:                      654052
Total generated tokens:                  655360
Request throughput (req/s):              3.72
Output token throughput (tok/s):         3805.72
Total Token throughput (tok/s):          7603.85
---------------Time to First Token----------------
Mean TTFT (ms):                          1452.02
Median TTFT (ms):                        1063.62
P99 TTFT (ms):                           5508.88
-----Time per Output Token (excl. 1st token)------
Mean TPOT (ms):                          32.21
Median TPOT (ms):                        32.42
P99 TPOT (ms):                           33.67
---------------Inter-token Latency----------------
Mean ITL (ms):                           32.21
Median ITL (ms):                         28.71
P99 ITL (ms):                            325.90
==================================================
vllm ({'pretrained': 'nvidia/Llama-3.3-70B-Instruct-FP4', 'kv_cache_dtype': 'fp8', 'tensor_parallel_size': 1, 'max_model_len': 2048, 'trust_remote_code': True}), gen_kwargs: (temperature=0.0), limit: 500.0, num_fewshot: 5, batch_size: 200
|Tasks|Version|     Filter     |n-shot|  Metric   |   |Value|   |Stderr|
|-----|------:|----------------|-----:|-----------|---|----:|---|-----:|
|gsm8k|      3|flexible-extract|     5|exact_match|↑  |0.932|±  |0.0113|
|     |       |strict-match    |     5|exact_match|↑  |0.842|±  |0.0163|

enable fusion:
============ Serving Benchmark Result ============
Successful requests:                     640
Maximum request concurrency:             128
Benchmark duration (s):                  170.02
Total input tokens:                      654052
Total generated tokens:                  655360
Request throughput (req/s):              3.76
Output token throughput (tok/s):         3854.49
Total Token throughput (tok/s):          7701.30
---------------Time to First Token----------------
Mean TTFT (ms):                          1430.17
Median TTFT (ms):                        1036.29
P99 TTFT (ms):                           5577.83
-----Time per Output Token (excl. 1st token)------
Mean TPOT (ms):                          31.80
Median TPOT (ms):                        32.01
P99 TPOT (ms):                           33.28
---------------Inter-token Latency----------------
Mean ITL (ms):                           31.80
Median ITL (ms):                         28.41
P99 ITL (ms):                            337.28
==================================================
vllm ({'pretrained': 'nvidia/Llama-3.3-70B-Instruct-FP4', 'kv_cache_dtype': 'fp8', 'tensor_parallel_size': 1, 'compilation_config': {'custom_ops': ['+silu_and_mul'], 'pass_config': {'enable_fusion': True, 'enable_noop': True}}, 'max_model_len': 2048, 'trust_remote_code': True}), gen_kwargs: (temperature=0.0), limit: 500.0, num_fewshot: 5, batch_size: 200
|Tasks|Version|     Filter     |n-shot|  Metric   |   |Value|   |Stderr|
|-----|------:|----------------|-----:|-----------|---|----:|---|-----:|
|gsm8k|      3|flexible-extract|     5|exact_match|↑  |0.926|±  |0.0117|
|     |       |strict-match    |     5|exact_match|↑  |0.852|±  |0.0159|
```



---
<details>
<summary> Essential Elements of an Effective PR Description Checklist </summary>

- [ ] The purpose of the PR, such as "Fix some issue (link existing issues this PR will resolve)".
- [ ] The test plan, such as providing test command.
- [ ] The test results, such as pasting the results comparison before and after, or e2e results
- [ ] (Optional) The necessary documentation update, such as updating `supported_models.md` and `examples` for a new model.
- [ ] (Optional) Release notes update. If your change is user facing, please update the release notes draft in the [Google Doc](https://docs.google.com/document/d/1YyVqrgX4gHTtrstbq8oWUImOyPCKSGnJ7xtTpmXzlRs/edit?tab=t.0).
</details>