@elvischenv 2025-08-27
Look like it is failed to create tensor on L4:
```
[2025-08-27T13:21:10Z] /usr/local/lib/python3.12/dist-packages/vllm/compilation/pass_manager.py:66: in configure
[2025-08-27T13:21:10Z]     self.passes += [ActivationQuantFusionPass(config)]
[2025-08-27T13:21:10Z] /usr/local/lib/python3.12/dist-packages/vllm/compilation/activation_quant_fusion.py:170: in __init__
[2025-08-27T13:21:10Z]     pattern_silu_mul_fp8.register(self.patterns)
[2025-08-27T13:21:10Z] /usr/local/lib/python3.12/dist-packages/vllm/compilation/activation_quant_fusion.py:100: in register
[2025-08-27T13:21:10Z]     self.empty_quant(5, 4),  # result
[2025-08-27T13:21:10Z] _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _
[2025-08-27T13:21:10Z]
[2025-08-27T13:21:10Z] self = <vllm.compilation.activation_quant_fusion.SiluMulFp8StaticQuantPattern object at 0x7f5cec68be00>
[2025-08-27T13:21:10Z] args = (5, 4), kwargs = {'device': 'cuda', 'dtype': torch.float8_e4m3fn}
[2025-08-27T13:21:10Z]
[2025-08-27T13:21:10Z]     def empty_quant(self, *args, **kwargs):
[2025-08-27T13:21:10Z]         kwargs = {'dtype': self.quant_dtype, 'device': "cuda", **kwargs}
[2025-08-27T13:21:10Z] >       return torch.empty(*args, **kwargs)
[2025-08-27T13:21:10Z] E       RuntimeError: CUDA error: no kernel image is available for execution on the device
```
I got a L4 locally and tried creating tensors and it worked. Is the failure related to the driver or something else for the L4 in CI? cc @ProExpertProg @mgoin 
```
$ nvidia-smi
Wed Aug 27 14:59:25 2025
+-----------------------------------------------------------------------------------------+
| NVIDIA-SMI 580.65.06              Driver Version: 580.65.06      CUDA Version: 13.0     |
+-----------------------------------------+------------------------+----------------------+
| GPU  Name                 Persistence-M | Bus-Id          Disp.A | Volatile Uncorr. ECC |
| Fan  Temp   Perf          Pwr:Usage/Cap |           Memory-Usage | GPU-Util  Compute M. |
|                                         |                        |               MIG M. |
|=========================================+========================+======================|
|   0  NVIDIA L4                      On  |   00000000:C1:00.0 Off |                    0 |
| N/A   42C    P8             16W /   72W |       0MiB /  23034MiB |      0%      Default |
|                                         |                        |                  N/A |
+-----------------------------------------+------------------------+----------------------+

+-----------------------------------------------------------------------------------------+
| Processes:                                                                              |
|  GPU   GI   CI              PID   Type   Process name                        GPU Memory |
|        ID   ID                                                               Usage      |
|=========================================================================================|
|  No running processes found                                                             |
+-----------------------------------------------------------------------------------------+
$ pip list | grep torch
torch                             2.7.1+cu128
torchaudio                        2.7.1+cu128
torchvision                       0.22.1+cu128
$ python -c "import torch; a = torch.ones((5,4), device='cuda', dtype=torch.float8_e4m3fn); print(a)"
tensor([[1., 1., 1., 1.],
        [1., 1., 1., 1.],
        [1., 1., 1., 1.],
        [1., 1., 1., 1.],
        [1., 1., 1., 1.]], device='cuda:0', dtype=torch.float8_e4m3fn)
```